Added a basic singleplayer campaign as an example... needs to be extended and a good level order needs to be planned.
No menu support yet. The commmand to start a singleplayer game is
"singleplayer". It will jump to the SP level you played last. If you win a
level, you advance to the next level. If you lose, you'll play the same level
again.

Badly needs a sound for "you've won" and "you've lost".



git-svn-id: svn://svn.icculus.org/nexuiz/trunk/data@1325 f962a42d-fe04-0410-a3ab-8c8b0445ebaa